telegram_tech_news/requirements.txt

feedparser==6.0.10
requests==2.31.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
matplotlib==3.8.2
flask==3.0.0
python-dotenv==1.0.0
pydub==0.25.1
soundfile==0.12.1
APScheduler==3.10.4
gunicorn==21.2.0
